refactor:Utility 생성하여 코드 리펙토링 진행
