+++ src/main/java/kr/codesquad/Prizes_Info.java
@@ -1,8 +1,10 @@
 package kr.codesquad;
 
 public enum Prizes_Info {
+
     THREE(5000) {
         public void print_message(int flag) {
+
             System.out.printf("3개 일치 (5000원) -%d개\n", flag);
         }
     },
